chore: update maintain scripts

+++ scripts/publish.sh
@@ -1,11 +1,21 @@
 #! /usr/bin/bash
 
 version=$(grep '^version = "' pack.toml | sed -e 's/version = "//' | sed -e 's/"//')
+if git tag | grep -q "$version"; then
+    echo "$version" is already tagged
+    exit
+fi
+changelog=$(git log "$(git tag | head -n1)"..HEAD --oneline --decorate=no --abbrev)
 
 rm -- *.mrpack *.zip
 
 scripts/update-credits.sh
 packwiz refresh
 
-git add .
-git commit -S -s -m "release: $version"
+git commit -uall -a -S -s --allow-empty -m "release: $version\n\n$changelog"
+git tag -S "$version"
+git push
+git push --tags
+
+scripts/sync-readme.sh
+CHANGELOG="$changelog" scripts/upload-to-mr.sh
